pyproject.toml
[project]
name = "gradcamfaith"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.4.0",
    "datasets>=3.5.0",
    "diffusers>=0.32.2",
    "einops>=0.8.1",
    "ipykernel>=6.29.5",
    "kaggle>=1.7.4.5",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "quantus>=0.5.3",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "seaborn>=0.13.2",
    "timm>=1.0.15",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.49.0",
]
